+++ apps/console/src/features/applications/components/wizard/help/generic-minimal-wizard-form-help.tsx
@@ -17,7 +17,7 @@
  */
 
 import { TestableComponentInterface } from "@wso2is/core/models";
-import { Heading, MessageWithIcon } from "@wso2is/react-components";
+import { Heading, Message } from "@wso2is/react-components";
 import React, { FunctionComponent, ReactElement } from "react";
 import { Divider } from "semantic-ui-react";
 import { ApplicationManagementConstants } from "../../../constants";
@@ -89,7 +89,7 @@ export const GenericMinimalWizardFormHelp: FunctionComponent<GenericMinimalWizar
                                         the
                                         application.
                                     </p>
-                                    <MessageWithIcon
+                                    <Message
                                         type="info"
                                         content={
                                             <>
