+++ apps/console/src/features/applications/components/wizard/oauth-protocol-settings-wizard-form.tsx
@@ -19,7 +19,7 @@
 import { TestableComponentInterface } from "@wso2is/core/models";
 import { URLUtils } from "@wso2is/core/utils";
 import { Field, FormValue, Forms } from "@wso2is/forms";
-import { ContentLoader, Hint, LinkButton, MessageWithIcon, URLInput } from "@wso2is/react-components";
+import { ContentLoader, Hint, LinkButton, Message, URLInput } from "@wso2is/react-components";
 import intersection from "lodash-es/intersection";
 import isEmpty from "lodash-es/isEmpty";
 import React, { FunctionComponent, ReactElement, useEffect, useState } from "react";
@@ -549,7 +549,7 @@ export const OauthProtocolSettingsWizardForm: FunctionComponent<OAuthProtocolSet
                                         showMoreContent={ t("common:showMore") }
                                     />
                                     { (callBackURLFromTemplate) && (
-                                        <MessageWithIcon
+                                        <Message
                                              type="info"
                                              content={
                                                  <>
