+++ apps/console/src/features/core/components/upload-certificate.tsx
@@ -18,7 +18,7 @@
 
 import { Certificate, TestableComponentInterface } from "@wso2is/core/models";
 import { CertificateManagementUtils } from "@wso2is/core/utils";
-import { GenericIcon, MessageWithIcon } from "@wso2is/react-components";
+import { GenericIcon, Message } from "@wso2is/react-components";
 import { KJUR, X509 } from "jsrsasign";
 import * as forge from "node-forge";
 import React, { FunctionComponent, ReactElement, useEffect, useRef, useState } from "react";
@@ -613,7 +613,7 @@ export const UploadCertificate: FunctionComponent<UploadCertificatePropsInterfac
 
             {
                 (fileError || certEmpty) && (
-                    <MessageWithIcon
+                    <Message
                         type="error"
                         data-testid={ `${ testId }-error-message` }
                         content={
