+++ apps/console/src/features/identity-providers/data/identity-provider-templates/templates/saml-identity-provider/saml-idp-wizard-file-based-help.tsx
@@ -22,7 +22,7 @@ import {
     CopyInputField,
     DocumentationLink,
     Heading,
-    MessageWithIcon,
+    Message,
     useDocumentation
 } from "@wso2is/react-components";
 import React, { FunctionComponent, ReactElement } from "react";
@@ -53,7 +53,7 @@ const SAMLIdPWizardFileBasedHelp: FunctionComponent<Props> = (props: Props): Rea
 
     return (
         <div data-testid={ testId }>
-            <MessageWithIcon
+            <Message
                 type="info"
                 content={
                     <>
