+++ modules/form/src/components/field-dropdown.tsx
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-import { Hint, MessageWithIcon } from "@wso2is/react-components";
+import { Hint, Message } from "@wso2is/react-components";
 import React, { ReactElement } from "react";
 import { Field as FinalFormField } from "react-final-form";
 import { SelectAdapter } from "./adapters";
@@ -51,7 +51,7 @@ export const FieldDropdown = (props: FieldDropdownPropsInterface): ReactElement 
         switch (props.message.type) {
             case "info":
                 return (
-                    <MessageWithIcon
+                    <Message
                         type={ props.message.type }
                         content={ props.message.content }
                         header={ props.message.header }
@@ -59,7 +59,7 @@ export const FieldDropdown = (props: FieldDropdownPropsInterface): ReactElement 
                 );
             case "warning":
                 return (
-                    <MessageWithIcon
+                    <Message
                         type={ props.message.type }
                         content={ props.message.content }
                         header={ props.message.header }
@@ -67,7 +67,7 @@ export const FieldDropdown = (props: FieldDropdownPropsInterface): ReactElement 
                 );
             case "error":
                 return (
-                    <MessageWithIcon
+                    <Message
                         type={ props.message.type }
                         content={ props.message.content }
                         header={ props.message.header }
