+++ modules/form/src/components/field-input.tsx
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-import { Hint, MessageWithIcon } from "@wso2is/react-components";
+import { Hint, Message } from "@wso2is/react-components";
 import React, { ReactElement } from "react";
 import { Field as FinalFormField } from "react-final-form";
 import { CopyFieldAdapter, PasswordFieldAdapter, TextFieldAdapter } from "./adapters";
@@ -146,7 +146,7 @@ export const FieldInput = (props: FieldInputPropsInterface): ReactElement => {
         switch (props.message.type) {
             case "info":
                 return (
-                    <MessageWithIcon
+                    <Message
                         type={ props.message.type }
                         content={ props.message.content }
                         header={ props.message.header }
@@ -154,7 +154,7 @@ export const FieldInput = (props: FieldInputPropsInterface): ReactElement => {
                 );
             case "warning":
                 return (
-                    <MessageWithIcon
+                    <Message
                         type={ props.message.type }
                         content={ props.message.content }
                         header={ props.message.header }
@@ -162,7 +162,7 @@ export const FieldInput = (props: FieldInputPropsInterface): ReactElement => {
                 );
             case "error":
                 return (
-                    <MessageWithIcon
+                    <Message
                         type={ props.message.type }
                         content={ props.message.content }
                         header={ props.message.header }
